unify relative timestamp functions
There are way too many places where we are declaring timestamp functions (e.g.
usec, ustime, mstime). We should only have one implementation.
