+++ src/quicklist.c
@@ -32,6 +32,7 @@
 #include "quicklist.h"
 #include "zmalloc.h"
 #include "ziplist.h"
+#include "time.h"
 #include "util.h" /* for ll2string */
 #include "lzf.h"
 
@@ -1459,20 +1460,6 @@ static void ql_info(quicklist *ql) {
     UNUSED(ql);
 #endif
 }
-
-/* Return the UNIX time in microseconds */
-static long long ustime(void) {
-    struct timeval tv;
-    long long ust;
-
-    gettimeofday(&tv, NULL);
-    ust = ((long long)tv.tv_sec) * 1000000;
-    ust += tv.tv_usec;
-    return ust;
-}
-
-/* Return the UNIX time in milliseconds */
-static long long mstime(void) { return ustime() / 1000; }
 
 /* Iterate over an entire quicklist.
  * Print the list if 'print' == 1.
